繰り返し処理構文を式に
